requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.19.1
fsspec==2025.7.0
hf-xet==1.1.8
huggingface-hub==0.34.4
idna==3.10
numpy==2.3.2
packaging==25.0
PyJWT==2.10.1
PyMySQL==1.1.1
python-decouple==3.8
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
safetensors==0.6.2
sqlparse==0.5.3
tokenizers==0.15.2
tqdm==4.67.1
transformers==4.35.2
typing_extensions==4.15.0
urllib3==2.5.0
